UPDATE: show value of ec version --short in output
This commit addresses HACBS-2018 and adds the same output as executing
`ec version --short` in the report output for `ec validate image` and
`ec validate definition`.
